GenerateOutput/requirements.txt
# Do not include azure-functions-worker in this file
# The Python Worker is managed by the Azure Functions platform
# Manually managing azure-functions-worker may cause unexpected issues

azure-functions
azure-storage-blob
azure-cosmos
gremlinpython
azure-ai-formrecognizer
openai
azure-core
azure-cosmos
azure-search-documents
langchain-openai
azure-identity